delete publisher check book

+++ eBookStoreClient/Pages/Publishers/Delete.cshtml.cs
@@ -3,9 +3,11 @@ using eBookStoreClient.Constants;
 using eBookStoreClient.Utilities;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using System.Collections.Generic;
 using System.Net;
 using System.Net.Http;
 using System.Text.Json;
+using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 
 namespace eBookStoreClient.Pages.Publishers
@@ -24,6 +26,8 @@ namespace eBookStoreClient.Pages.Publishers
 
         [TempData]
         public int PublisherId { get; set; }
+
+        public string HasBookMessage { get; set; }
 
         public async Task<IActionResult> OnGetAsync(int? id)
         {
@@ -68,11 +72,23 @@ namespace eBookStoreClient.Pages.Publishers
             try
             {
                 HttpClient httpClient = SessionHelper.GetHttpClient(HttpContext.Session, sessionStorage);
-                HttpResponseMessage response = await httpClient.DeleteAsync($"{Endpoints.Publishers}/{PublisherId}");
+                HttpResponseMessage response = await httpClient.GetAsync($"{Endpoints.Books}?$filter=PublisherId eq {PublisherId}");
                 HttpContent content = response.Content;
                 if (response.StatusCode == HttpStatusCode.OK)
                 {
-                    return RedirectToPage(PageRoute.Publishers);
+                    var books = JsonSerializer.Deserialize<Books>(await content.ReadAsStringAsync(), SerializerOptions.CaseInsensitive).List;
+                    if (books.Count > 0)
+                    {
+                        HasBookMessage = "Unable to delete publisher. This publisher is assigned to book(s).";
+                        return await OnGetAsync(PublisherId);
+                    }
+                    httpClient = SessionHelper.GetHttpClient(HttpContext.Session, sessionStorage);
+                    response = await httpClient.DeleteAsync($"{Endpoints.Publishers}/{PublisherId}");
+                    content = response.Content;
+                    if (response.StatusCode == HttpStatusCode.OK)
+                    {
+                        return RedirectToPage(PageRoute.Publishers);
+                    }
                 }
                 return await OnGetAsync(Publisher.PublisherId);
             }
@@ -81,5 +97,11 @@ namespace eBookStoreClient.Pages.Publishers
             }
             return RedirectToPage(PageRoute.Publishers);
         }
+    }
+
+    class Books
+    {
+        [JsonPropertyName("value")]
+        public List<Book> List { get; set; }
     }
 }
